comment: Diary 조회 주석 추가
- 더 빠른 조회를 위해 고민을 했지만 지식을 더 쌓고 다시 도전하기로 함
- 고민의 흔적을 주석으로 남김

+++ src/main/java/com/phcworld/repository/board/dto/DiarySelectDto.java
@@ -25,7 +25,7 @@ public class DiarySelectDto {
 
     private Long countOfAnswers;
 
-    private Long countOfGood;
+    private Integer countOfGood;
 
     private LocalDateTime updateDate;
